+++ src/main/kotlin/quickfix/models/User.kt
@@ -14,7 +14,7 @@ abstract class User : Id {
     abstract var gender : Gender
     abstract var address : Address
 
-    override var id: Int = -1
+    override var id: Long = -1
 
     companion object {
         const val EDAD_REQUERIDA = 18
